Refactor handlers to own file
